moved tensor search tests to correct dir
